requirements.txt
livereload
jinja2
pyyaml
requests
beautifulsoup4